Fix bugs in monthsBetweenDtgs and yearsBetweenDtgs
The existing definition of yearsBetweenDtgs is sensitive to leap-year
issues, as it tries to handle years as durations.  For example, the
number of years between 2009-03-01 and 2008-03-01 is resolved to be 0.

This commit redefines yearsBetweenDtgs as just monthsBetweenDtgs/12.
It also replaces the definition of monthsBetweenDtgs with one that
is *mostly* based around naively differencing the month values, but
(hopefully) does the right thing when the later date is at an earlier
day of the month (or earlier time within the same day) than the earlier
date.

Added quite a few tests to try to exercise as many corner cases as I
could think of, but any further ideas (or counter-examples!) would be
nice.