Fixed bugs in monthsBetweenDtgs and yearsBetweenDtgs (#19)
* Fix bugs in monthsBetweenDtgs and yearsBetweenDtgs

The existing definition of yearsBetweenDtgs is sensitive to leap-year
issues, as it tries to handle years as durations.  For example, the
number of years between 2009-03-01 and 2008-03-01 is resolved to be 0.

This commit redefines yearsBetweenDtgs as just monthsBetweenDtgs/12.
It also replaces the definition of monthsBetweenDtgs with one that
is *mostly* based around naively differencing the month values, but
(hopefully) does the right thing when the later date is at an earlier
day of the month (or earlier time within the same day) than the earlier
date.

Added quite a few tests to try to exercise as many corner cases as I
could think of, but any further ideas (or counter-examples!) would be
nice.

* Rewrote tests as parameterised class

I've also removed a few tests I put in that on hindsight seemed
redundant, and added a few I missed (such as getting the months between
a DTG and itself, or trying a timespan long enough to hopefully cover
all of the Gregorian calendar leap year rules!)

* Added postconditions for betweenDtgs functions

* Changed addMonths, subtractMonths to nat domain

This means we can no longer define subtracting x months as subtracting
-x months.

This required a bit of changing to handle the modular arithmetic
on months in a way that was consistent between both functions.  Hoping
I've not broken anything on the way.

* Fixed unsound postcondition wrt addMonths clipping

* Pass for consistency

* Fixed accidentally weakened postconditions

* Added add/subtractYears and basic tests

* Addressed comments from Simon

Also possibly one or two out of band from Alessandro.